+++ build-tests/tests/fail/net_without_uds_missing_uds.rs
@@ -1,0 +1,4 @@
+use build_tests::tokio_net::udp;
+
+fn main() {}
+
