+++ build-tests/tests/fail/net_without_uds_missing_uds.stderr
@@ -1,0 +1,7 @@
+error[E0432]: unresolved import `build_tests::tokio_net::udp`
+ --> $DIR/net_without_uds_missing_uds.rs:1:5
+  |
+1 | use build_tests::tokio_net::udp;
+  |     ^^^^^^^^^^^^^^^^^^^^^^^^^^^ no `udp` in `tokio_net`
+
+For more information about this error, try `rustc --explain E0432`.
