+++ tokio-net/Cargo.toml
@@ -32,6 +32,13 @@ udp = [
     "futures-sink-preview",
     "futures-util-preview",
 ]
+uds = [
+    "bytes",
+    "mio-uds",
+    "futures-util-preview",
+    "iovec",
+    "libc",
+]
 
 [dependencies]
 tokio-codec = { version = "=0.2.0-alpha.1", path = "../tokio-codec" }
@@ -54,6 +61,11 @@ bytes = { version = "0.4", optional = true }
 futures-sink-preview = { version = "=0.3.0-alpha.18", optional = true }
 futures-util-preview = { version = "=0.3.0-alpha.18", optional = true }
 iovec = { version = "0.1", optional = true }
+
+# UDS
+[target.'cfg(unix)'.dependencies]
+mio-uds = { version = "0.6.5", optional = true }
+libc = { version = "0.2.42", optional = true }
 
 [dev-dependencies]
 tokio = { version = "=0.2.0-alpha.1", path = "../tokio" }
