+++ tokio-net/src/uds/datagram.rs
@@ -1,5 +1,5 @@
-use tokio_net::driver::Handle;
-use tokio_net::util::PollEvented;
+use crate::driver::Handle;
+use crate::util::PollEvented;
 
 use futures_core::ready;
 use futures_util::future::poll_fn;
