--- tokio-uds/Cargo.toml
@@ -1,43 +1,0 @@
-[package]
-name = "tokio-uds"
-# When releasing to crates.io:
-# - Remove path dependencies
-# - Update html_root_url.
-# - Update doc url
-#   - Cargo.toml
-# - Update CHANGELOG.md.
-# - Create "v0.3.x" git tag.
-version = "0.3.0-alpha.1"
-edition = "2018"
-authors = ["Tokio Contributors <team@tokio.rs>"]
-license = "MIT"
-repository = "https://github.com/tokio-rs/tokio"
-homepage = "https://github.com/tokio-rs/tokio"
-documentation = "https://docs.rs/tokio-uds/0.3.0-alpha.1/tokio_uds/"
-description = """
-Unix Domain sockets for Tokio
-"""
-categories = ["asynchronous"]
-
-[features]
-async-traits = []
-
-[dependencies]
-tokio-codec = { version = "=0.2.0-alpha.1", path = "../tokio-codec" }
-tokio-net = { version = "=0.2.0-alpha.1", path = "../tokio-net" }
-tokio-io = { version = "=0.2.0-alpha.1", path = "../tokio-io" }
-
-bytes = "0.4.8"
-mio = "0.6.14"
-mio-uds = "0.6.5"
-futures-core-preview = "=0.3.0-alpha.18"
-futures-util-preview = "=0.3.0-alpha.18"
-iovec = "0.1.2"
-libc = "0.2.42"
-log = "0.4.2"
-
-[dev-dependencies]
-tokio = { version = "=0.2.0-alpha.1", path = "../tokio" }
-
-tempfile = "3"
-futures-preview = "=0.3.0-alpha.18"
