+++ tokio/src/net.rs
@@ -67,7 +67,7 @@ pub use self::udp::{UdpFramed, UdpSocket};
 pub mod unix {
     //! Unix domain socket bindings for `tokio` (only available on unix systems).
 
-    pub use tokio_uds::{split, UCred, UnixDatagram, UnixListener, UnixStream};
+    pub use tokio_net::uds::{split, UCred, UnixDatagram, UnixListener, UnixStream};
 }
 #[cfg(all(unix, feature = "uds"))]
 pub use self::unix::{UnixDatagram, UnixListener, UnixStream};
